restructure hexkit
